Added basic level loading

+++ src/Tile.cpp
@@ -3,5 +3,5 @@
 Tile::Tile(SDL_Renderer* renderer, std::string path, bool hasTransparency) :
 	Bitmap(renderer, path, hasTransparency)
 {
-	setSize(Vector2f(surface->w, surface->h));
+	setSize(Vector2f((float)surface->w, (float)surface->h));
 }
